Add files via upload
Compared to our UML Diagram from part 2.1, the new updated version is much more organized and readable. A few of the major changes between the two include creating the Adventurer and Creature as abstract classes, splitting each creature/adventurer into their own subclasses, adding in a compass, and how we formed it all into a general hierarchy. The overall move set and ending logic was shuffled around as well, but we found it necessary to add another class when we began coding how the rooms related to each other. That's why the compass was added, it allows for movement and how to find specific locations.